% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\docType{data}
\name{annotation_libraries}
\alias{annotation_libraries}
\title{Annotation libraries}
\format{
Each element of the list is an annotation library in the form of
a list of gene sets. The annotation libraries are a selection of the ones
provided by Enrichr (see link below).
}
\source{
\url{https://maayanlab.cloud/Enrichr/#stats}
}
\usage{
annotation_libraries
}
\description{
Annotation libraries
}
\keyword{datasets}
