% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{assess_xnnet_performance}
\alias{assess_xnnet_performance}
\title{Assessing performance of xnnet predictions given labels}
\usage{
assess_xnnet_performance(xnnet, xnnet_predictions, true_labels, metric = "AUC")
}
\arguments{
\item{xnnet}{xnnet object}

\item{xnnet_predictions}{xnnet predictions on a set of samples}

\item{true_labels}{true labels of predicted class}

\item{metric}{metric to evaluate the performance}
}
\description{
Assessing performance of xnnet predictions given labels
}
