% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{build_xnnet}
\alias{build_xnnet}
\title{Building xnnet}
\usage{
build_xnnet(
  X_train,
  y_train,
  annotation_libraries,
  n_input_nodes = 3,
  n_hidden_nodes = 4,
  n_unassigned = 3,
  number = 10,
  min_decay = 0.1,
  max_decay = 1
)
}
\arguments{
\item{X_train}{data matrix}

\item{y_train}{binary labels}

\item{annotation_libraries}{a list of annotation libraries. Each library is a list of gene sets
gene sets}
}
\description{
This function builds xnnet using a labeled training set and a list of annotation libraries
}
\examples{
data("GSE37250") #load Tubercolosis dataset
data("annotation_libraries")
GSE37250_split = train_test_split(GSE37250$X, GSE37250$y)
xnnet = build_xnnet(X_train = GSE37250_split$X_train, y_train = GSE37250_split$y_train,
annotation_libraries = annotation_libraries)
}
