% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{compute_activation}
\alias{compute_activation}
\title{Computing activation state of hidden nodes}
\usage{
compute_activation(xnnet_single, X, y, zscore_threshold = 3)
}
\arguments{
\item{xnnet_single}{output of build_xnnet}

\item{X}{matrix of transcriptional profiles (samples x genes)}

\item{y}{binay vector of labels for each sample}

\item{zscore_threshold}{threshold on zscores}
}
\description{
Computing activation state of hidden nodes
}
