% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{get_and_process_GSEA_results}
\alias{get_and_process_GSEA_results}
\title{Renormalize NES produced by GSEA}
\usage{
get_and_process_GSEA_results(annotation_libraries, limma_results)
}
\arguments{
\item{annotation_libraries}{list of gene annotation libraries}

\item{limma_results}{results of limma (output of get_limma_results)}
}
\description{
Renormalize NES produced by GSEA
}
