% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{get_GSEA_results}
\alias{get_GSEA_results}
\title{Gene Set Enrichment Analysis (GSEA)}
\usage{
get_GSEA_results(
  limma_results,
  annotation_library,
  min_term_size = 10,
  max_term_size = Inf,
  rank_metric = "logFC"
)
}
\arguments{
\item{limma_results}{out of Limma}

\item{annotation_library}{list of gene sets contained in a given annotation library}

\item{min_term_size}{min size of annotation term (default = 10)}

\item{max_term_size}{max size of annotation term (default = Inf)}

\item{rank_metric}{statistic used to rank genes for GSEA}
}
\description{
This function performs GSEA using the Limma output and an annotation library
}
