% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\docType{data}
\name{GSE37250}
\alias{GSE37250}
\title{Transcriptomics data from active vs latent tuberculosis patients}
\format{
Sample of a GEO dataset containing labeled transcriptional profiles from active and latent tuberculosis patients
\describe{
\item{X}{expression matrix with 50 samples (rows) and 31426 probes (columns)}
\item{y}{a vector of binary labels for each sample: 0 = latent tuberculosis, 1 = active tuberculosis}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE37250}
}
\usage{
GSE37250
}
\description{
Transcriptomics data from active vs latent tuberculosis patients
}
\keyword{datasets}
