% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_plotting_functions.R
\name{plot_xnnet}
\alias{plot_xnnet}
\title{Plot xxnet structure}
\usage{
plot_xnnet(xnnet, color = "gray70")
}
\arguments{
\item{xnnet}{output of build_xnnet}

\item{color}{color of input nodes and edges}
}
\description{
This function splits a dataset and corresponding
labels into a training and test set
}
