% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{plot_xnnet_heatmap}
\alias{plot_xnnet_heatmap}
\title{Plot xxnet results as a heatmap}
\usage{
plot_xnnet_heatmap(xnnet)
}
\arguments{
\item{xnnet}{output of build_xnnet}
}
\description{
This function plots xnnet as a heatmap with genes as rows and samples as columns.
Genes are grouped by the corresponding annotation and rows are grouped by sample class.
labels into a training and test set
}
