% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_plotting_functions.R
\name{plot_xnnet_input_genes}
\alias{plot_xnnet_input_genes}
\title{Plot xxnet results as a heatmap}
\usage{
plot_xnnet_input_genes(xnnet, hidden_node)
}
\arguments{
\item{xnnet}{output of build_xnnet}

\item{hidden_node}{hidden node index}
}
\description{
This function plots xnnet as a heatmap with genes as rows and samples as columns.
Genes are grouped by the corresponding annotation and rows are grouped by sample class.
labels into a training and test set
}
