% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{reduce_GSEA_results}
\alias{reduce_GSEA_results}
\title{Reducing GSEA results with Weighted Set Cover}
\usage{
reduce_GSEA_results(GSEA_results, topN = 10)
}
\arguments{
\item{GSEA_results}{GSEA results (output of get_GSEA_results)}

\item{topN}{The number of sets (or less when it completes early) to return.}
}
\value{
A list of \code{topSets} and \code{coverage}.
\describe{
\item{topSets}{A list of set IDs.}
\item{coverage}{The percentage of IDs covered in the top sets.}
}
}
\description{
Size constrained weighted set cover problem to find top N sets while
maximizing the coverage of all elements.
}
