% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{renormalize_NES}
\alias{renormalize_NES}
\title{Renormalize NES produced by GSEA}
\usage{
renormalize_NES(GSEA_results)
}
\arguments{
\item{GSEA_results}{output of get_GSEA_results}
}
\description{
Renormalize NES produced by GSEA
}
