% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{select_xnnet_nodes}
\alias{select_xnnet_nodes}
\title{Defining xnnet nodes}
\usage{
select_xnnet_nodes(
  reduced_GSEA_results,
  limma_results,
  n_input_nodes = 5,
  n_hidden_nodes = 4
)
}
\arguments{
\item{reduced_GSEA_results}{reduced GSEA results}

\item{limma_results}{Limma results}

\item{n_input_nodes}{number of input nodes (default = 5)}

\item{n_hidden_nodes}{number of input nodes (default = 4)}
}
\description{
This function selects the xnnet input and hidden nodes
}
