% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{threshold_scaled_data}
\alias{threshold_scaled_data}
\title{Predict sample labels using xnnet}
\usage{
threshold_scaled_data(scaled_data, z_threshold)
}
\arguments{
\item{xnnet}{xxnet object return by the function build_xnnet}

\item{X_test}{test data}
}
\description{
Predict sample labels using xnnet
}
