% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{train_test_split}
\alias{train_test_split}
\title{Training/test split}
\usage{
train_test_split(X, y, training_fraction = 0.7, rnd_seed = 666)
}
\arguments{
\item{X}{data matrix}

\item{y}{binary labels}

\item{training_fraction}{fraction of data for training}
}
\value{
A list containing \code{(X_train, y_train, X_test, y_test)}.
}
\description{
This function splits a dataset and corresponding
labels into a training and test set
}
\examples{
data("GSE37250") #load Tubercolosis dataset
GSE37250_split = train_test_split(GSE37250$X, GSE37250$y)
}
