% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{xnnet_predict}
\alias{xnnet_predict}
\title{Generating xnnet predictions on a set of samples}
\usage{
xnnet_predict(xnnet, X_test, z_threshold = 3)
}
\arguments{
\item{xnnet}{transcriptomics data with samples as rows and genes as features}

\item{X_test}{binary label for each sample in X}

\item{z_threshold}{fraction of data for training}
}
\description{
This function produces class probabilities on a set of samples
labels into a training and test set
}
