% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{assess_xnnet_performance}
\alias{assess_xnnet_performance}
\title{Assessing performance of xnnet predictions given labels}
\usage{
assess_xnnet_performance(xnnet, xnnet_predictions, true_labels)
}
\arguments{
\item{xnnet_predictions}{output of build_xnnet}

\item{true_labels}{hidden node index}
}
\description{
This function plots xnnet as a heatmap with genes as rows and samples as columns.
Genes are grouped by the corresponding annotation and rows are grouped by sample class.
labels into a training and test set
}
