% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{build_xnnet}
\alias{build_xnnet}
\title{Building xnnet}
\usage{
build_xnnet(
  X_train,
  y_train,
  annotation_libraries,
  n_input_nodes = 3,
  n_hidden_nodes = 4,
  n_unassigned = 3,
  number = 10,
  min_decay = 0.1,
  max_decay = 1
)
}
\arguments{
\item{X_train}{data matrix}

\item{y_train}{binary labels}

\item{annotation_libraries}{a list of annotation libraries. Each library is a list of gene sets
gene sets}

\item{n_input_nodes}{maximum number of input nodes per hidden nodes}

\item{n_hidden_nodes}{number of hidden nodes in the network}

\item{n_unassigned}{number of input genes not assigned to a specific gene set}

\item{number}{number of bootstrap samples for network training}

\item{min_decay}{minimum value of the decay parameter for regularization}

\item{max_decay}{maximum value of the decay parameter for regularization}
}
\value{
A list containing an xnnet object for each annotation library
}
\description{
This function builds xnnet using a labeled training set of transcriptional profiles,
and a list of annotation libraries
}
\examples{
data("GSE37250") #load Tubercolosis dataset
data("annotation_libraries")
GSE37250_split = train_test_split(GSE37250$X, GSE37250$y)
xnnet = build_xnnet(X_train = GSE37250_split$X_train, y_train = GSE37250_split$y_train,
annotation_libraries = annotation_libraries)
}
