% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{get_limma_results}
\alias{get_limma_results}
\title{Limma differential analysis}
\usage{
get_limma_results(X_train, y_train, adj_pval_cutoff = 0.05)
}
\arguments{
\item{X_train}{data matrix}

\item{y_train}{binary labels}

\item{adj_pval_cutoff}{fraction of data for training (default = 0.05)}
}
\description{
This function performs differential gene expression with Limma on
the training set
}
