% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnnet_functions_v0.1.R
\name{xnnet_predict}
\alias{xnnet_predict}
\title{Predict sample labels using xnnet}
\usage{
xnnet_predict(xnnet, X_test)
}
\arguments{
\item{xnnet}{xxnet object return by the function build_xnnet}

\item{X_test}{test data}
}
\description{
Predict sample labels using xnnet
}
